ipu3: fix stripe calculation
fix a couple of calculations that I believe should not be adjusting the width
by 1. This patch resolves 8-byte alignment errors when transforming to
I420 mode by fixing the fact that spliting a 1280 width by 2 should result
in a 640/640 split instead of 632/648. When the split is un-even it can
result in the U and/or V offsets not being 8-byte aligned which causes a
Y and UV mis-alignment in the 2nd strip of the image which is clearly visible.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>
Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>