imx7d-var-som: Limit the max cpu freq. to 1 GHz
According to documents AN11957, AN5334 and IMX7DCEC from NXP,
running an iMX7 part other than MCIMX7D2DVK12SD and
MCIMX7D2DVM12SD (which have restricted specs and a baseline
feature-set), at a core freq. of 1.2 GHz is possible for
evaluation, but might reduce the lifetime of the part.

This patch limits the max cpu freq. to 1 GHz
(actually 996 MHz), in accordace with the part's specification.

Signed-off-by: Eran Matityahu <eran.m@variscite.com>