can: mcp25xxfd: optionally enable SPI crc commands for reads
Allow the (optional) use of SPI CRC read commands that can get enabled
via a module parameter.

In the future it may become necessary to enable this for register reads
(or at least some specific register reads) as there is a HW bug that
may result in glitches on the spi bus.

Possible variants here are:
* just retry the read
* do not use some fifos (7, 15, 23, 31) at all
* test for bit flips of specific bits and retest CRC

--
Changelog:
  V7 -> V8: added in V8

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>
Signed-off-by: Eran Matityahu <eran.m@variscite.com>