imx8mq-var-dart: move M4 reserved RAM at 0x7e000000
Currently the M4 RAM is located at 0x80000000: this location is
available only for SoMs with 2GB of RAM or more.
For some reason, the M4 cannot access RAM below address 0x60000000.
This means that M4 cannot access RAM on SOMs with 512MB of RAM.
Moving the area at 0x7e000000 will allow M4 accessing the RAM on
SoMs with 1GB of RAM or more.
This change must match an equivalent change in freertos repository.
Pls refer to related changes in freertos branch
   mcuxpresso_sdk_2.5.x-var01