imx8mm-var-dart: add Cortex-M4 related device tree
-moved M4 reserved RAM at 0x7e000000
On NXP EVK, the M4 RAM is located at 0x80000000: this location is
available only for SoMs with 2GB of RAM or more.
For some reason, the M4 cannot access RAM below address 0x60000000.
This means that M4 cannot access RAM on SOMs with 512MB of RAM.
Moving the area at 0x7e000000 allows M4 accessing the RAM on
SoMs with 1GB of RAM or more.
This change must match linker files settings in freertos repository.
Pls refer to related settings in freertos branch
  mcuxpresso_sdk_2.5.x-var01