gpio-mxc: fix issue multiply setting pads wakeup on the same gpio bank
This patch fixes the issue in the case of we need to configure more gpio lines
to pad wakeup on the same gpio bank. Without the fix only the first pad configured
as pad wakeup works fine. For the others gpio lines the below error appears:
  - gpio-mxc xxxxxxxx.gpio: set pad wakeup failed, ret -17

Example of device tree with multiply pad wakeup on the same gpio bank:
&lsio_gpio5 {
	/* Enable wakeup-source from two gpio pins*/
	pad-wakeup-num = <2>;  /* total pad wakeup number in gpio5 */

	/*
	 Each triple means:
	 gpio-key pin id
	 trigger type
	 gpio pin
	*/
	pad-wakeup = <IMX8QM_USDHC1_DATA4 SC_PAD_WAKEUP_LOW_LVL 19>, <IMX8QM_USDHC1_STROBE SC_PAD_WAKEUP_LOW_LVL 23>;
};