imx8mq-var-dart: Remove "hard-wired" PCIE0 controller property
The "hard-wired" DTS property of NXP PCIE controller driver is used
to mark PCIE controllers that have hard-wired PCIE devices on the
carrier board (i.e. WIFI module on i.MX8MQ EVK). This is not the case
on DT8MCustomBoard, which comes with mini-PCIE connector attched to
the PCIE0 controller.