Revert "Revert "gpio: pca953x: Disable PM routines""
This reverts commit 0243069daad804780a1255e6927efeba82046bb7.
With DT8MCustomBoard v2.x the PCIe RST pin is connected to one of
the pins of the TCA6408 GPIO expander.
During the power up sequences, specifically for DART-MX8M, the
kernel don't correctly register the PCIe as a consumer of TCA6408.
During the suspend, the TCA6408 is suspended before the PCIe
(it should be the other way round).
As side effect, the PCIe is resumed before the TCA6408, but trying
to toggle the RST pin, nothing happens and the system hangs.
As workaround, disable the PM routines.