imx8mp-var-dart: dts: add aliases for ecspi instances
The spi-imx device driver defers waiting for SDMA resources to be
initialized during the boot. This behavior can cause swapped SPI
indexes as the indexes are chosen accordingly depending on which
SPI instance successfully passes first. Aliasing the SPI devices
ensures finding a static index mapping in the Userspace in any case.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>