imx8qxp-var-som: disable unused gpio banks
Derived from f9dbb0e

Without this patch the gpio banks are enumerated starting from gpio2 due
to the gpio0 and gpio1 are assigned to gpio0_mipi_csi0 and gpio0_mipi_csi1.
This make the wifi and BT not working.
Also aliases for the gpio0_mipi_csi0 and gpio0_mipi_csi1 don't fix the issue.

Signed-off-by: Diego Dorta <diego.d@variscite.com>