net: mdio: force deassert MDIO reset signal
When the reset gpio is defined within the node of the device tree
describing the PHY, the reset is initialized and managed only after
calling the fwnode_mdiobus_phy_device_register function.
However, before calling it, the MDIO communication is checked by the
get_phy_device function.
When this happen and the reset GPIO was somehow previously set down,
the get_phy_device function fails, preventing the PHY detection.
These changes force the deassert of the MDIO reset signal before
checking the MDIO channel, but also free the gpio descriptor to
allow managing it later.

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>