net: phy: adin: Add flag to allow enabling of 1000base-t retrain
The PHY contains a retrain function which if enabled, retrains the
1000base-t link when the receiver status becomes not okay. If disabled,
the link will simply drop. This retrain function should be enabled in
order to be IEEE compliant, but is disabled by default on the PHY due to
FLD being enabled by default.

Add a device tree flag which can be used to enable the function on the
PHY when desired.

Signed-off-by: Ken Sloat <ken.s@variscite.com>
(cherry picked from commit 7c29de903ef150587d56aae25a26217aa69c5db5)