dts: imx8mn-var-som-symphony-m7: add clocks property
In moving from Kernel 5.15.71 to 6.1.x, NXP added the clocks property
to the imx8mn-cm7 node.

NXP appears to have rebased on this upstream, so there is no single
commit to observe this isolated change in Linux - however the difference
can be seen by doing a diff of the kernel versions themselves.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>