dt-bindings: net: adin: Document bindings for 1000BASE-T retrain
The ADI PHY contains a "retrain" feature which will allow a 1000BASE-T
link to retrain automatically if the receiver status becomes not ok.
This feature should be enabled to be IEEE compliant but is disabled on
the PHY by default.

Document the new device tree binding "adi,1000base-t-retrain-en" which
allows the user to optionally enable this feature.

Signed-off-by: Ken Sloat <ken.s@variscite.com>

Signed-off-by: Harshesh Valera <harshesh.v@variscite.com>