imx93-var-som: add support for iw612 802.15.4 module
VAR-SOM-MX93 variants ordered with the WBE option do not have the
resistive touch panel controller assembled. The interrupt and SPI chip
select are instead routed to the iw612 module.

This patch adds a new device tree that disables the touch controller
and adds a spidev device required for OpenThread to manage 802.15.4
in userspace.

Signed-off-by: Nate Drude <nate.d@variscite.com>