dts: imx8qm-var-som: fix the cm4 node build warning
The new implementation of_get_cpu_hwid() in imx_rproc.c parsing the
"reg" in cm4 node, but "reg" keywords refer to reg_format and in this
node it conflicts with address-cells/size-cells rule and causes build
warning. Put the node in a cm4 node and set the correct
address-cells/size-cells value to fix the issue.

For further details, see:
nxp-imx/linux-imx@409efbc4edd5

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>