dts: imx8qm-var-som: rename resource table reserved mem dt name
Derived from commit 6f9d3441569e.

In moving from Kernel 5.15.71 to 6.1.x, NXP changed the node name of the
resource table for the M4/rproc from "rsc_table" to "rsc-table." Since
the imx-rproc driver references this node by name, this change is
mandatory.

NXP appears to have rebased on this upstream, so there is no single
commit to observe this isolated change in Linux - however the difference
can be seen by doing a diff of the kernel versions themselves.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>