dts: imx8qm-var-som: align unit address
Derived from commit b56ad201e386.

In moving from Kernel 5.15.71 to 6.1.x, NXP aligned on some nodes
the unit address dropping the hexadecimal prefix.

NXP appears to have rebased on this upstream, so there is no single
commit to observe this isolated change in Linux - however the difference
can be seen by doing a diff of the kernel versions themselves.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>