net: phy: at803x: configure rgmii voltage in config init
The RGMII voltage should also be configured during the config init
phase. Therefore the patch consolidate the logic to set the voltage,
and add it to the config_init function.

This fix an issue where the RGMII voltage is not set on boot or is
lost after a suspend and resume sequence.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>
(cherry picked from commit ed296d0dbfd48a04a51068de2ca9feacd65d9b58)