net:phy: mxl-8611x: configure rgmii rx clock delay
The mxl8611x supports a fixed RGMII Rx Clock Delay. The delay is 2 ns
for 125 Mhz or 8 ns for 2.5/25 MHz. The initial setting is defined
by pin strapping.

The delay should be disabled for rgmii-txid and rgmii phy-modes.

Upstream-Status: Pending (This patch should be pushed upstream once
the driver is available upstream)

Cherry-picked from commit 3004e8891486.

Signed-off-by: Nate Drude <nate.d@variscite.com>