ov5640: fix pixel rate and link frequency calculation
The ratio between link frequency and pixel clock must not depend on the 'mipi_div' prescaler value,
which only determines how the clocks are generated.
The link frequency only depends on pixel clock, data format and number of lanes.
The pixel clock period in the relevant register is always in unit of 0.5 ns regardless of
MIPI prescaler value.

Signed-off-by: Marco Contenti <marco.c@variscite.com>