imx8mp-var-dart, imx8mn-var-som: fix issue due to OV5640 camera driver changes
The pixel rate calculation in the previous commit caused a regression in imx8mp and imx8mn
where low-resolution modes do not work anymore with the current csis-hs-settle value.
Changing the value from 13 to 6 fixes the issue and supports all resolutions.
As an alternative we could have changed the MIPI driver in a way similar to commit 74f8158
in which this value is calculated as (f - 5)/45 with f = link frequency in MHz. No source
is given for this formula.
Table 13 in AN13857 contains values to be set vs. frequency ranges. Since the NXP kernel
release supplies a single value for a range much wider than those in the table, such values
cannot be considered mandatory for each given range.

Signed-off-by: Marco Contenti <marco.c@variscite.com>