ARM: configs: imx_v7_var_defconfig: disable FW loader fallback
When the file-system is based on systemd, if a specific firmware file
is missing, this configuration forces systemd to keep waiting until the
firmware timeout expire (60 seconds).
While generally this is not a problem, the WiFi driver for broadcom
chips always try loading a file with extension ".clm_blob" which is
actually available only for a subset of chips (like CYW4343W, used by
Sterling LWB).
When the SoM is mounted with Sterling LWB5 (based on CYW43353), this
firmware file is not available and the firmware loading process stuck
until the firmware timeout expires.
Generally speaking, CONFIG_FW_LOADER_USER_HELPER_FALLBACK usage is
deprecated in almost all recent Linux distros.

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>