imx93-var-som: move eqos phy reset to ethphy0 node
GPIO1_I07 controls only the phy reset. The phy regulator is always
enabled according to the assembly of the SOM.

Using reset-gpios to manage the phy reset allows the MDIO framework
to manage the phy. The reset will be asserted when the interface is down,
reducing power consumption during suspend.

Signed-off-by: Nate Drude <nate.d@variscite.com>