imx8qxp-var-som: Add missing VPU node to the device tree
Add the missing VPU node to the VAR-SOM-MX8X device tree to
ensure proper initialization and functionality of the video
processing unit. This update aligns with the NXP's mek device
tree and enables the core VPU settings that were previously
missing.

The update addresses issues related to video playback and
GStreamer plugin availability that were observed due to the
lack of proper VPU configuration.

Signed-off-by: Diego Dorta <diego.d@variscite.com>