arm64: imx8_var_defconfig: Disable fw loader user helper
The brcmfmac driver may fail with the FW_LOADER_USER_HELPER option enabled due
to delays in firmware loading, additional dependencies on user-space processes,
or compatibility issues with the driver's initialization flow. To avoid these
failures, it is recommended to disable FW_LOADER_USER_HELPER if the system is
already able to load the firmware correctly via the kernel directly, ensuring
that the firmware is loaded more quickly and directly without relying on
external processes.

Signed-off-by: Luan Rafael Carneiro <luan.r@variscite.com>