arm: dts: imx6ul-imx6ull: Migrate PHY reset from FEC node to PHY node
The Scarthgap release already uses the modern and recommended approach
of handling PHY resets inside the PHY node itself (ethernet-phy@x),
using the properties: reset-gpios, reset-assert-us, and reset-deassert-us.

In the Kirkstone release, the reset was still defined under the MAC
controller (FEC) using deprecated properties: phy-reset-gpios, and
phy-reset-duration. This approach has a main issue, it is marked as
deprecated in the official device tree binding: fsl,fec.yaml file.

To align Kirkstone with Scarthgap and follow current best practices,
all the reset logic to the PHY nodes were migrate. This enables full
control over reset timing, removes deprecated syntax, and guarantees
future compatibility with modern PHYs such as MaxLinear or others
with stricter timing.

The following properties were removed from &fec1 and &fec2:

- phy-reset-gpios
- phy-reset-duration

And the following were added to each ethernet-phy@x:

- reset-gpios = <...>;
- reset-names = "phy";
- reset-assert-us = <10000>;       // 10 ms reset pulse
- reset-deassert-us = <100000>;    // 100 ms wait before MDIO

This change is fully compliant with:

- Documentation/devicetree/bindings/net/ethernet-phy.yaml
- Documentation/devicetree/bindings/net/fsl,fec.yaml

The GPIO values for the reset-gios was copied from this patch:

- varigit/linux-imx@ea4a04dbba38 ARM: dts: imx6ul: Fix reset pins for ethernet PHYs on Variscite SOMS

Signed-off-by: Diego Dorta <diego.d@variscite.com>