ARM: dts: i2c clock frequency, fix typo
The I2C DT property for the clock frequency is called clock-frequency,
not clock_frequency.

This has no effect on most I2C devices, as the default I2C clock
frequency of 100000 would be chosen anyway, if 'clock-frequency' is not
provided.

On var-dart-6ulcustomboard, this will set i2c1 to 100kHz. Due to the
typo, i2c1 has always operated at 100kHz, even though it looked like it
should have operated at 400kHz. This commit does not modify the factual
clock frequency, but it documents it correctly and allows customers to
modify it if desired.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>