imx8mn-var-som: Enable Ethernet PHY regulator early at boot
Commits 8ec378a and 8ec378a reverted the logic that, if the
Ethernet PHY reset was asserted in U-Boot, the kernel would deassert it
before scanning MDIO for PHYs.

With that logic removed, the PHY now requires its regulator GPIO to be
asserted as early as possible during boot so that the MDIO bus can
detect and properly initialize the PHY.

Add the 'regulator-boot-on' property to enable the PHY regulator at an
early stage.

Based on: caf3c8f

Signed-off-by: Andre Morishita <andre.m@variscite.com>