imx8mm-var-dart: Enable Ethernet PHY regulator early at boot
Commits 8ec378a8c5ac and 1e565912bb07 reverted the logic that, if the
Ethernet PHY reset was asserted in U-Boot, the kernel would deassert it
before scanning MDIO for PHYs.

With that logic removed, the PHY now requires its regulator GPIO to be
asserted as early as possible during boot so that the MDIO bus can
detect and properly initialize the PHY.

Add the 'regulator-boot-on' property to enable the PHY regulator at an
early stage.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>
(cherry picked from commit 271a3ab957195ab574ebf559951d9492df434dde)